+++ src/aqua/cli/diagnostic_config.py
@@ -109,6 +109,11 @@ diagnostic_config = {
             'target_path': 'diagnostics/ocean3d'
         },
         {
+            'config_file': 'config_ocean_stratification.yaml',
+            'source_path': 'config/diagnostics/ocean3d',
+            'target_path': 'diagnostics/ocean3d'
+        },
+        {
             'config_file': 'config_ocean_drift.yaml',
             'source_path': 'config/diagnostics/ocean3d',
             'target_path': 'diagnostics/ocean3d'